+++ src/net/mod.rs
@@ -1,3 +1,3 @@
-pub mod request;
+pub mod http;
 pub mod tcp;
 pub mod util;
